+++ src/analysis/slices.rs
@@ -317,7 +317,7 @@ impl FunctionSlices {
             StepResult::Continue | StepResult::LoadStore { .. } => {
                 let next_address = ins_addr + 4;
                 // If we already visited the next address, connect the blocks and end
-                if executor.visited(section.address as u32, next_address)
+                if executor.visited(section.address, next_address)
                     || self.blocks.contains_key(&next_address)
                 {
                     self.blocks.insert(block_start, Some(next_address));
